Organize repository files
